requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
pycairo==1.25.1
PyGObject==3.46.0
python-dotenv==1.0.1
requests==2.31.0
setuptools==68.2.2
urllib3==2.2.0
websockets==12.0
Werkzeug==3.0.1
wheel==0.41.2
